check quote is not finish and continue quote still have bugs
