+++ srcs/ms_implement.c
@@ -6,7 +6,7 @@
 /*   By: sthitiku <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/24 23:12:53 by sthitiku          #+#    #+#             */
-/*   Updated: 2022/12/24 23:21:39 by sthitiku         ###   ########.fr       */
+/*   Updated: 2023/01/14 23:54:27 by sthitiku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,4 +33,41 @@ int	echo(char **args, int newline)
 	if (newline == 1)
 		ft_putchar_fd('\n', 1);
 	return (1);
+}
+
+char	*get_pwd()
+{
+	return (getcwd(NULL, MAXPATHLEN));
+}
+
+/*
+ * Change the current working directory.
+ *	- If the path is a relative or absolute path, it will be used directly.
+ *	- If the path is related to HOME (`cd ~` or `cd`), it will be expanded to absolute path.
+ *
+ * @param path_str The path to the new working directory.
+ * @return 0 if the path is valid, -1 otherwise.
+*/
+int	ms_cd(char *path_str)
+{
+	int		status_code;
+	char	*new_path;
+
+	new_path = NULL;
+	if (path_str[0] == '~' || path_str[0] == '\0')
+	{
+		new_path = malloc(sizeof(char) * (ft_strlen(getenv("HOME")) + 1 + ft_strlen(path_str) - 1));
+		ft_strlcpy(new_path, getenv("HOME"), ft_strlen(getenv("HOME")) + 1);
+		ft_strlcpy(&new_path[ft_strlen(new_path)], &path_str[1], ft_strlen(path_str));
+	}
+	else
+	{
+		new_path = malloc(sizeof(char) * (ft_strlen(path_str) + 1));
+		ft_strlcpy(new_path, path_str, ft_strlen(path_str) + 1);
+	}
+	status_code = access(new_path, F_OK);
+	if (!status_code)
+		chdir(new_path);
+	free(new_path);
+	return (status_code);
 }
